+++ tests/test-dirs/type-enclosing/underscore-ids.t
@@ -294,6 +294,30 @@ We try several places in the identifier to check the result stability
     "value": [
       {
         "start": {
+          "line": 5,
+          "col": 9
+        },
+        "end": {
+          "line": 5,
+          "col": 12
+        },
+        "type": "int",
+        "tail": "no"
+      },
+      {
+        "start": {
+          "line": 5,
+          "col": 4
+        },
+        "end": {
+          "line": 5,
+          "col": 12
+        },
+        "type": "int option",
+        "tail": "no"
+      },
+      {
+        "start": {
           "line": 2,
           "col": 18
         },
@@ -301,7 +325,19 @@ We try several places in the identifier to check the result stability
           "line": 5,
           "col": 17
         },
-        "type": "int",
+        "type": "int option -> int",
+        "tail": "no"
+      },
+      {
+        "start": {
+          "line": 2,
+          "col": 10
+        },
+        "end": {
+          "line": 5,
+          "col": 17
+        },
+        "type": "'a",
         "tail": "no"
       },
       {
@@ -313,7 +349,7 @@ We try several places in the identifier to check the result stability
           "line": 5,
           "col": 17
         },
-        "type": "'a -> int",
+        "type": "'a -> 'b",
         "tail": "no"
       }
     ],